refactor: BlackJackGame에서 Dto 의존성 분리
